[BUGFIX] Open a new tab if window name in backend is `newTYPO3frontendWindow`
If a preview is opened via the TYPO3 backend, the opened window has the
name `newTYPO3frontendWindow` to ensure only one preview window is open.
However, if somebody changes in such window into the TYPO3 backend, the
window name is not changed, causing that such preview links are opened
in the same window again.

This patch resets the window name before a preview is opened to ensure
that previews are always opened in a new window.

Resolves: #89805
Releases: master, 9.5
Change-Id: Ib4b31e78da21befc27c50ea3d49762f2a81fe029
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62491
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>