[TASK] Restructure nightly plans in stages a 50 jobs
The rearranged stages structure will prevent the host
machines and also gerrit from being overwhelmed with a
high amount of running jobs all at once.
Instead, the equal size of each stage keeps the load
on an average height, but it will increase the total
runtime of a plan. We prefer stability over quick runs
in nightlies.

Resolves: #89865
Releases: master, 9.5, 8.7
Change-Id: Ib915053e83d087007242c05886619faf8d7cee0f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62546
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62555