[TASK] Set lastLoginProvider cookie only if custom provider used
This patch changes the login flow so it only sets the
be_lastLoginProvider cookie if the provider that
authenticated the login, is different from the primary
login provider configured in TYPO3.

The result is that the cookie is not set, as it is not
necessary, when the primary provider is used.

Resolves: #89877
Releases: master
Change-Id: I20c9163b88396e1e6efdd257d890fd62b4fa9205
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62556
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>