[BUGFIX] Send HTTP headers with PSR-7 response
Current Backend User Authentication sends header()
which is not appended to the headers of the PSR-7
response and cannot be tested or validated properly.

This is mainly due to legacy reasons as AbstractUserAuthentication
sends these headers.

When using TYPO3 in a scenario to do sub-requests
within one PHP process, it is not possible to properly
evaluate these Response headers.

To enable this, the BackendUserAuthenticator middlewares
now apply the headers to the Response object.

Resolves: #89911
Releases: master
Change-Id: Id22ca1a65e52f101d3775fbe79ea0ef1622e9fa9
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62593
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Tobi Kretschmann <tobi@tobishome.de>
Tested-by: Susanne Moog <look@susi.dev>
Reviewed-by: Jörg Bösche <typo3@joergboesche.de>
Reviewed-by: Sascha Rademacher <sascha.rademacher+typo3@gmail.com>
Reviewed-by: Tobi Kretschmann <tobi@tobishome.de>
Reviewed-by: Susanne Moog <look@susi.dev>