[BUGFIX] Apply thumbscrews to sqlserver
sqlserver is a special race that welcomes high memory servers as its
habitat. However, we are short on resources and need to offer a very
small amount of RAM for a living which already acts as server cruelty.

While we're on it, it gets a nice bed to put some data on the RAM disk
in the hope its job gets done at least quicker.

Resolves: #89935
Releases: master, 9.5, 8.7
Change-Id: Ic06f4b7071ed1d16ec9d0cd5a0a57f05dde171f6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62614
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>