[TASK] Simplify checks on Import about allowed filenames
In TYPO3 v10, BasicFileUtility is not used by DataHandler anymore,
and never initialized with other file permissions than "allow
any file ending", which is why the simple check on the fileDenyPattern
can be used directly in Import functionality.

The next patch would then allow to clean up BasicFileUtility
where the methods which belong to "internal_type=file" are not in use
anymore.

Resolves: #89941
Releases: master
Change-Id: I58a28a1cd1f6d284d241479a324c877a2824e7bb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62627
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Susanne Moog <look@susi.dev>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Susanne Moog <look@susi.dev>