[TASK] Thin out SystemEnvironmentBuilder
This a pre-patch to clean up the functionality of the SystemEnvironmentBuilder,
which was originally introduced as part of the Bootstrap logic in 6.0.

However, a few cross-concerns can be cut in TYPO3 v10:
- Calling the deprecated GeneralUtility::presetApplicationContext() is not
needed anymore, as it can be populated on-demand when the deprecated
GeneralUtility::getApplicationContext() method is called via Environment API.
This way, the ApplicationContext initialization can be put directly in the
Environment API initialization call.
- The static "isFunctionDisabled" method with a nasty runtime cache is only
used in one place in backend, so does not belong to "building system environment",
and is therefore removed.
- The now unused "exitWithMessage" protected method is removed,
removing a dependency to the HttpUtility class.

Resolves: #89943
Releases: master
Change-Id: I31156b1a1ded306d99bcf2d51de43bc919a0b3e0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62628
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Tested-by: Susanne Moog <look@susi.dev>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Susanne Moog <look@susi.dev>