[BUGFIX] PopulatePageSlugs return the most recent slug
The populate page slugs functionality in de upgrade wizard does not
work correct for pages with multiple title adjustments.

Realurl keeps for every page title change a record in
tx_realurl_pathdata.

In order to get the latest change a order should be added to the
query in getSuggestedSlugs.

Resolves: #89964
Releases: master, 9.5
Change-Id: I4b7fad14fc559e6c9e8a4fb6f6a2a112a290d7e2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62795
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>