[BUGFIX] Test failure due to erroneous usage of uniqueId() function
The function in testing-framework provides a string that is target to be
used e.g. for HTML tags, but the BackendUserAuthenticationTest uses it
to produce a random userId, which needs to be an integer.
If the randomly produced string contains not only numbers, the cast
operation converts the uid to 0 and therefor prevents a DB query
from execution. This prevention results in a not consumed mock object,
which in turn causes a integration test failure.

Resolves: #90018
Releases: master, 9.5
Change-Id: I559041591a600a04da86e8b4a85cf5e6dd176475
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62773
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>