[FEATURE] Native MountPoint support for Site Handling
In order to link to mountpoints with Site Handling, the RootlineUtility
needs to receive the MountPoint parameter, to correctly
deal with mountpoint-related subpages.

Mount Points are based on the assumption that the "context" (the original
site environment) is kept. For this to work, the slug of the mount page (doktype=7)
is prefixed to the URL, but the common prefix of the subpages is trimmed
with the value of the mounted page (= pointer record).

The MP parameter is then added to PageArguments within the PageArguments
as "routeArgument" (= safe and clean argument) where TYPO3 is dealing
with this feature again in the same fashion it as before.

Various side-effects when dealing with mount points from other domains
still exist (= different language setup, or non-existing sites).

Feature Set:
* Multi-language setup (= when language setup is the same) with slugs
* Recursive mount points
* No MP parameter available in URLs anymore (at all)
* Multi-site setup (= when language setup is the same)

If a subpage of a mount page does not inherit the slug of the mounted page,
then the slug of the subpage is added in full afterwards.

Resolves: #86331
Resolves: #87473
Resolves: #89039
Releases: master, 9.5
Change-Id: I58f41eb325a07cc0c4a0dfeab1164eb8c58c7314
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62933
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Marco Huber <mail@marco-huber.de>
Tested-by: Daniel Siepmann <coding@daniel-siepmann.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Marco Huber <mail@marco-huber.de>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Benni Mack <benni@typo3.org>