[BUGFIX] Access module uses DataHandler for updating page
The permission access module allows for AJAX calls to dynamically
updating page permissions on the fly, where DataHandler is used.

When using the form to do a more customizable change,
plain SQL was used previously to update the pages. This circumvented
the change to also update any translated pages (which DataHandler does
by default).

This wasn't a problem until TYPO3 v9.0, where pages_language_overlay
was merged with "pages", where the permissions of translated pages
should reflect the changes of the default language page.

The change updates the PermissionController->updateAction
to call DataHandler for the changes as well.

Resolves: #89838
Resolves: #88311
Releases: master, 9.5
Change-Id: I9e56b279ddcbb43ee7bb11b330c288b12a0e7439
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62930
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Jörg Bösche <typo3@joergboesche.de>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Jörg Bösche <typo3@joergboesche.de>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>