[FEATURE] Introduce modal actions
Action buttons in modals created by the `TYPO3/CMS/Backend/Modal`
module may now make use of `ImmediateAction` and `DeferredAction`.

As an alternative to the existing `trigger` option, the new option
`action` may be used with an instance of the previously mentioned
modules.

A similar feature has been introduced with #89061 for notifications.

This patch also implements the feature in the recycler module.

Resolves: #90168
Releases: master
Change-Id: I3fde2ced31e6bd2b0779524c0aeead17a2edaa49
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63008
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>