[TASK] Dissolve PersistenceDelegate for persisted mappers
The intention of PersistenceDelegate was to provide a generic API to
resolve or generate route components. However the implementation did
not provide any additional behavior. It just was used as structural
pattern which did not provide any further advantages.

That's the reason why PersistenceDelegate gets dissolved and its
database connection invocation is moved to its corresponding callers
PersistedAliasMapper and PersistedPatternMapper.

For backward-compatibility reasons previous (protected) implementation
is still kept in legacy layer for those two route aspects.

Resolves: #90218
Releases: master, 9.5
Change-Id: I49eb58372c139c9b8274593efa2892a43eb81508
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63048
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>