[BUGFIX] Use list module search in page module
The page module search bar now redirects to the list module,
where all functionality is given, and all records on this
page (not just tt_content) are found, streamlining
it with the list module logic.

Editors without permission for the list module do not see
the search in the page module (similar behaviour to LiveSearch).

All code of the page layout classes related to search is removed.
As this is non-public API, the removal is done without any
deprecations necessary.

Resolves: #90235
Releases: master
Change-Id: Ia23c931c5a009e8ffe9a1a56219fb922d63e44bf
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63063
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>