[FEATURE] Dashboard for TYPO3
A dashboard is introduced into TYPO3 to show the most important
information to the current logged in user.

Every user with access to this backend module can now have one
or more personal dashboards. Each dashboard can contain several
widgets. Which widgets and in which order the widgets are shown
is up to the users themselves.

Resolves: #90333
Releases: master
Change-Id: I964a52846ec9a1c901baa89508b9df0caecdf81f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63139
Tested-by: Koen Wouters <koen.wouters@maxserv.com>
Tested-by: Riny van Tiggelen <info@online-gamer.nl>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Björn Jacob <bjoern.jacob@tritum.de>
Reviewed-by: Koen Wouters <koen.wouters@maxserv.com>
Reviewed-by: Riny van Tiggelen <info@online-gamer.nl>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Björn Jacob <bjoern.jacob@tritum.de>