[TASK] Remove unused code in New CE Wizard / PagePositionMap
There are a few places where unused options can be simplified,
and the dependency on the PagePositionMap class can be placed in
one dedicated method.

Several guard clauses are useless due to method parameter types,
and are cleaned up.

This is a pre-patch to bring more functionality into New Content
Element Wizard.

Resolves: #90366
Releases: master
Change-Id: I912efbe25f6ff79670f404d7f44dbad0993295db
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63211
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>