[BUGFIX] Change no_follow, no_index labels to their new inverted labels
The fields no_follow and no_index were inverted some time ago, but still
retained their previous non-inverted labels in TCA. These were not
actually shown when editing a page, but could be read in by other parts
of the backend.
This lead to misleading labels in both the "Display information" window
as well as the "Info -> Pagetree Overview" module.

Resolves: #88812
Releases: master, 9.5
Change-Id: Ie1bddeae20f0ffb3958bb07dab78d68dc68eff74
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63312
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>