[BUGFIX] Render exception for backend preview as error
Prior only an warning was logged, and a fallback rendering jumped in.
With this change an error is rendered in addition to logging. No
fallback will be triggered anymore.

This should raise awareness and ease integration for integrators.
Please note that already broken elements might now show an error
instead of fallback preview.

This is only activated for admin users and enabled BE|debug system
setting.

Resolves: #90427
Releases: master
Change-Id: I1f24a1e775ba4f12917ed26dc29b0ea78ab81a1f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63318
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Björn Jacob <bjoern.jacob@tritum.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Eßl <indy.essl@gmail.com>
Reviewed-by: Ingo Fabbri <ingo.fabbri@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Christian Eßl <indy.essl@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>