[BUGFIX] Allow multiple mountpoints with the same slug
When having a multisite setup using the same slugs pointing to the same
or other different mounted pages, the PageSlugCandidateProvider will
always fetch the information of the first page found in the database
which complies with the slug. In some cases this will be the right page,
but in other cases will be a page which is not part of the requested
website. In the last case this will result in a 404.

This patch fixes this by checking if the slug is from the site being
called.

Resolves: #90469
Releases: master, 9.5
Change-Id: I333bfdfeb37ffa13b86697c4597cbda6e315b590
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64057
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Patrick Broens <patrick@patrickbroens.nl>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Patrick Broens <patrick@patrickbroens.nl>
Reviewed-by: Benni Mack <benni@typo3.org>