[TASK] Introduce Argon2PasswordHashInterface
This is a followup patch for the introduction of Argon2id.
https://review.typo3.org/c/Packages/TYPO3.CMS/+/63077

That commit introduced access to class constants that are
only available through late static binding. This approach
does not follow a contract and uses a bit of a magical
approach instead. It's hard to read for the user, cannot
be type checked and it is possibly unstable and may produce
unforeseen issues when inheriting AbstractArgon2PasswordHash
without defining said class constants.

To have a clean architecture with a contract driven approach
an interface is introduced which child classes of
AbstractArgon2PasswordHash have to implement. This ensures
that the algorithm in AbstractArgon2PasswordHash can
reliably fetch necessary information from the child classes.

Releases: master
Resolves: #90508
Change-Id: I950da09762bcc859b5c03ad8ffeeefe46af2ed02
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63391
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Susanne Moog <look@susi.dev>