[BUGFIX] Fix Exception caused by empty staticPageArguments
When 'Index Records immediately when saved?' is enabled in indexing
configurations for records, then on saving the record the
string 'null' is used for this indexingconfiguration and passed to
SearchController::preparePageLink(), which expects an array and
therefore throws an exception.

This commit fixes this by checking if
$this->conf['staticPageArguments'] is an array and only then json_decode
that, else it sets the value to null.

Resolves: #89816
Related: #86994
Releases: master, 9.5
Change-Id: I96d3fb963ecb508bc536967afece064ad12bb989
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63414
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63448