[BUGFIX] Prevent TypeError in AssetCollector related ViewHelpers
The two new ViewHelpers `CssViewHelper` and `ScriptViewHelper`
automatically render the content of their body if no source file
is supplied in the ViewHelpers argument. However if the body is
empty, NULL is passed to the AssetCollectors add inline code functions.
This leads to a TypeError as these parameters are type hinted.

Therefore the body content is now checked before
the methods get called to prevent passing NULL or an empty string.

Resolves: #90539
Releases: master
Change-Id: I80feb9551f50e6ab8ac56740b605c4bbbfbe3069
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63437
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Jens Ulrich <jens.ulrich@snk.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Jens Ulrich <jens.ulrich@snk.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>