[TASK] use strict type comparison for isValueSelected
\TYPO3\CMS\Fluid\ViewHelpers\Form\SelectViewHelper::isSelected
uses a strict type comparison but the OptionViewHelper does not.

This resulted in funny results when the values were convertable to int
like this:
var_dump(in_array(1, ['1-2'])); // prints bool(true)

The value is now converted to a string, just like in the
SelectViewHelper, and then strictly compared.

Releases: master
Resolves: #90540
Change-Id: I6c28316ca306f5188141a0222b6f09a1d1fc108e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63439
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>