[BUGFIX] Ensure presence of record uid for postModifiers of slugs
When recreating a slug via the datahandler, the full record is present
in the hook, whereas when recreating the same slug via the ajax request
only a sub-set of values is sent. To ensure postModifiers are able
to recreate and modify slugs based on the original record, the handling
has been changed to always provide the record uid for further processing.

Resolves: #89862
Releases: master, 9.5
Change-Id: Ib0d0b86be1c544d44c1b7420619b21afd5650988
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63447
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>