[TASK] Setup missing constants in phpstan bootstrap file
This patch introduces a bootstrap file for phpstan to define
missing constants phpstan would report otherwise.

A dedicated config file phpstan.typo3.neon has been
introduced because there are plans to release a seperate
configuration package that not only the core uses but
which is also useful for TYPO3 users that test their
codebase with phpstan.

Releases: master
Resolves: #90575
Change-Id: I1c5bce558c882f9dd002e480d78405770bdf8cc0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63480
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Tested-by: Susanne Moog <look@susi.dev>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Susanne Moog <look@susi.dev>