[BUGFIX] Only allow scalar values in AssetCollector
Since AssetCollector needs to be serialized,
or to be more precise its state need to be cached as string,
we must avoid that state is stored that isn't serializable.

Therefore we adapt the API to reflect this and change
the API calls to not add non serializable objects.

With this change we also revert the previous USER_INT fix
that bound the AssetCollector unnecessarily to PageRenderer
and store it serialized in cache (the deserialization was
already part of the code of the commit that introduced
the AssetCollector).

Reverts: #90537
Resolves: #90565
Releases: master
Change-Id: I59f10b608e5f270aaacde63bcd347116910aac7a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63488
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>