[TASK] Replace $.ajax() in most of EXT:backend
This patch replaces $.ajax() in EXT:backend either with our AJAX API
or with native XMLHttpRequest. The latter is required in the DragUploader
as we need to listen to the progress event which isn't supported by
fetch().

This patch skips Storage/Persistent, as this one needs some refactoring
due to the usage of synchronous(!) AJAX requests.

Resolves: #90597
Releases: master
Change-Id: Ifdf41c133929740f3e766ae992848ab99ca1f283
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63502
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Tested-by: Susanne Moog <look@susi.dev>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Susanne Moog <look@susi.dev>