[BUGFIX] Show scrollbar earlier in pagetree on larger pagetrees
On larger pagetrees, the svg pagetree cut off too many elements at
the bottom and the scrollbar was activated much too late. This was
because the calculated css height of the pagetree could become larger
than the height of the html body itself.
Also redraw the height when the page doktype icon list gets updated.

Resolves: #90555
Resolves: #89193
Releases: master, 9.5
Change-Id: Ice8c46110e2e8c92469a3d2ab12a7744bd6fc7de
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63511
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Riccardo De Contardi <erredeco@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>