[BUGFIX] Do not assign tab items if their content is empty
In a TabsContainer, if all the fields of a tab are hidden to the editor
(by disabling them in PageTS), the tab would still become assigned in
the 'items' fluid variable. The empty, but invisible, tab would then
still be considered for determining the active tab in the view. If the
first tab was disabled in this way, the editor would see no active tab
at all.
To prevent this from happening, the tab elements are now never assigned,
if their content is completely empty.

Resolves: #89094
Releases: master, 9.5
Change-Id: I3cbe98e7e25b7827fdc68583387343f5f0cb64ca
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63512
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>