[TASK] Remove bootstrap-slider usages
This patch removes all occurrences of bootstrap-slider in Core with the
native `range` input type.
To ensure a consistent styling across all platforms, new CSS is in place.
However, we can't recover the same styling as used with
the bootstrap-slider either due to inconsistent browser support of
selectors or missing features.

In the `ValueSlider` module, which is only used by FormEngine, any usage
of jQuery has been removed as well.

Resolves: #90649
Releases: master
Change-Id: Ic2e0cb7aa9305159c33f14130d3fe6e62d2fd1da
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63523
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>