[BUGFIX] Include Page Read Permissions to PageTreeController
In order to reduce the amount of data fetched, the PageTree loaded
for editors will only fetch pages that the editor actually sees, reducing
the actual amount of pages to fetch.

If a pagetree has 80.000 pages, but the editor can only see 50, the
tree only loads these 50 pages.

Resolves: #90880
Releases: master, 9.5
Change-Id: I57985484ace07fbfb919573351c210b291697ae3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63983
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Susanne Moog <look@susi.dev>
Reviewed-by: Susanne Moog <look@susi.dev>